+++ src/main/java/com/irontechspace/dynamicdq/configurator/core/repository/IConfigRepository.java
@@ -1,0 +1,25 @@
+package com.irontechspace.dynamicdq.configurator.core.repository;
+
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface IConfigRepository<T> {
+
+    List<T> getAll();
+
+    List<T> getAll(NamedParameterJdbcTemplate jdbcTemplate);
+
+    void savePosition(T config);
+
+    void savePosition(NamedParameterJdbcTemplate jdbcTemplate, T config);
+
+    T save(T config);
+
+    T save(NamedParameterJdbcTemplate jdbcTemplate, T config);
+
+    void delete(UUID configId);
+
+    void delete(NamedParameterJdbcTemplate jdbcTemplate, UUID configId);
+}