+++ src/main/java/com/irontechspace/dynamicdq/executor/ExecutorType.java
@@ -1,0 +1,21 @@
+package com.irontechspace.dynamicdq.executor;
+
+// Перечисление типы задачи
+public enum ExecutorType {
+    flat,
+    hierarchical,
+    count,
+    object,
+    sql,
+    sqlCount,
+    save,
+    queue,
+    equal,
+    not_equal,
+    greater_equal,
+    less_equal,
+    greater,
+    less,
+    log,
+    branch,
+}