+++ src/main/java/com/irontechspace/dynamicdq/executor/file/model/File.java
@@ -1,0 +1,12 @@
+package com.irontechspace.dynamicdq.executor.file.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class File  {
+    private String name;
+    private String absolutePath;
+    private byte[] content;
+}