+++ src/main/java/com/irontechspace/dynamicdq/executor/query/model/TypeQuery.java
@@ -1,0 +1,9 @@
+package com.irontechspace.dynamicdq.executor.query.model;
+
+public enum TypeQuery {
+    COUNT,
+    TABLE,
+    OBJECT,
+    SQL,
+    SQL_COUNT
+}