+++ src/main/java/com/irontechspace/dynamicdq/model/TypeQuery.java
@@ -1,9 +1,0 @@
-package com.irontechspace.dynamicdq.model;
-
-public enum TypeQuery {
-    COUNT,
-    TABLE,
-    OBJECT,
-    SQL,
-    SQL_COUNT
-}