+++ src/main/java/com/irontechspace/dynamicdq/rabbit/model/RabbitTaskOutputType.java
@@ -1,0 +1,3 @@
+package com.irontechspace.dynamicdq.rabbit.model;
+
+public enum RabbitTaskOutputType { database, queue }