+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.37
+
+- Refactor struct project
+- Add rabbit tasks
+
 ## 0.0.32
 
 - Add `json` type data for SET data